some refactor for node.js app
