+++ nodejs-to-cratedb/.eslintrc
