+++ nodejs-to-cratedb/settings/twitterAuthExample.js
@@ -1,0 +1,8 @@
+module.exports = {
+  authSetting: {
+    consumer_key: 'your consumer key',
+    consumer_secret: 'your consumer secret',
+    token: 'your auth token',
+    token_secret: 'your secret token'
+  }
+};
